requirements.txt
ccxt
anthropic
pandas
pandas-ta
sqlalchemy
psycopg2-binary
python-dotenv
aiohttp
websockets
APScheduler
requests
feedparser
google-generativeai
